+++ CHANGELOG.md
@@ -11,10 +11,16 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 - :fire: Remove public getters from classes. Now should be access via property instead.
 - :fire: Remove Pascal Case in properties. Properties updated to be camel case only.
-- :fire: Remove deprecated classes under `ǸelsonMartell\Utilities` namespace.
+- :fire: Remove deprecated classes under `NelsonMartell\Utilities` namespace.
+- :bug: :fire: Fix problem in `NelsonMartell\StrictObject::compare()` for some types and make it more generic.
+- :new: `class`: `NelsonMartell\Extensions\Numbers`.
+- :new: `class`: `NelsonMartell\Extensions\Arrays`.
 
 
 ### :notebook: Development changes
+
+- :art: Improve doc blocks.
+- :up: Improve tests and increase the code coverage.
 
 
 ### :bookmark: More changes